+++ tests/scm/config/security/test_dns_security_profile.py
@@ -16,7 +16,7 @@ from scm.models.security.dns_security_profiles import (
     LogLevelEnum,
     PacketCaptureEnum,
 )
-from tests.factories import (
+from tests.test_factories.security import (
     BotnetDomainsFactory,
     DNSSecurityCategoryEntryFactory,
     DNSSecurityProfileCreateApiFactory,
