+++ tests/test_factories/security/__init__.py
@@ -14,6 +14,19 @@ from tests.test_factories.security.decryption_profile import (
     DecryptionProfileUpdateApiFactory,
     DecryptionProfileUpdateModelFactory,
 )
+from tests.test_factories.security.dns_security_profile import (
+    BotnetDomainsFactory,
+    DNSSecurityCategoryEntryFactory,
+    DNSSecurityProfileCreateApiFactory,
+    DNSSecurityProfileCreateModelFactory,
+    DNSSecurityProfileResponseFactory,
+    DNSSecurityProfileUpdateApiFactory,
+    DNSSecurityProfileUpdateModelFactory,
+    ListActionRequestFactory,
+    ListEntryBaseFactory,
+    SinkholeSettingsFactory,
+    WhitelistEntryFactory,
+)
 from tests.test_factories.security.url_categories import (
     URLCategoriesCreateModelFactory,
     URLCategoriesResponseModelFactory,
@@ -28,12 +41,23 @@ __all__ = [
     "AntiSpywareProfileResponseFactory",
     "AntiSpywareProfileUpdateApiFactory",
     "AntiSpywareProfileUpdateModelFactory",
+    "BotnetDomainsFactory",
     "DecryptionProfileCreateApiFactory",
     "DecryptionProfileUpdateApiFactory",
     "DecryptionProfileResponseFactory",
     "DecryptionProfileCreateModelFactory",
     "DecryptionProfileUpdateModelFactory",
+    "DNSSecurityCategoryEntryFactory",
+    "DNSSecurityProfileCreateApiFactory",
+    "DNSSecurityProfileCreateModelFactory",
+    "DNSSecurityProfileResponseFactory",
+    "DNSSecurityProfileUpdateApiFactory",
+    "DNSSecurityProfileUpdateModelFactory",
+    "ListActionRequestFactory",
+    "ListEntryBaseFactory",
+    "SinkholeSettingsFactory",
     "URLCategoriesCreateModelFactory",
     "URLCategoriesResponseModelFactory",
     "URLCategoriesUpdateModelFactory",
+    "WhitelistEntryFactory",
 ]
